+++ tests/examples/test_cartis_roberts.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
-from estimagic.examples.cartis_roberts import CARTIS_ROBERTS_PROBLEMS
-from estimagic.examples.cartis_roberts import get_start_points_bdvalues
+from estimagic.benchmarking.cartis_roberts import CARTIS_ROBERTS_PROBLEMS
+from estimagic.benchmarking.cartis_roberts import get_start_points_bdvalues
 from numpy.testing import assert_array_almost_equal
