+++ tests/examples/test_more_wild.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
-from estimagic.examples.more_wild import get_start_points_mancino
-from estimagic.examples.more_wild import MORE_WILD_PROBLEMS
+from estimagic.benchmarking.more_wild import get_start_points_mancino
+from estimagic.benchmarking.more_wild import MORE_WILD_PROBLEMS
 
 
 @pytest.mark.parametrize("name, specification", list(MORE_WILD_PROBLEMS.items()))
