+++ tests/examples/test_noise_distributions.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
 import pytest
-from estimagic.examples.benchmarking import _sample_from_distribution
-from estimagic.examples.noise_distributions import NOISE_DISTRIBUTIONS
+from estimagic.benchmarking.benchmarking import _sample_from_distribution
+from estimagic.benchmarking.noise_distributions import NOISE_DISTRIBUTIONS
 
 
 @pytest.mark.parametrize("distribution", NOISE_DISTRIBUTIONS)
