+++ tests/examples/test_process_benchmark_results.py
@@ -1,17 +1,17 @@
 import numpy as np
 import pandas as pd
 import pytest
-from estimagic.examples.process_benchmark_results import _clip_histories
-from estimagic.examples.process_benchmark_results import _find_first_converged
-from estimagic.examples.process_benchmark_results import (
+from estimagic.benchmarking.process_benchmark_results import _clip_histories
+from estimagic.benchmarking.process_benchmark_results import _find_first_converged
+from estimagic.benchmarking.process_benchmark_results import (
     _get_history_as_stacked_sr_from_results,
 )
-from estimagic.examples.process_benchmark_results import (
+from estimagic.benchmarking.process_benchmark_results import (
     _get_history_of_the_parameter_distance,
 )
-from estimagic.examples.process_benchmark_results import _make_history_monotone
-from estimagic.examples.process_benchmark_results import _normalize
-from estimagic.examples.process_benchmark_results import create_performance_df
+from estimagic.benchmarking.process_benchmark_results import _make_history_monotone
+from estimagic.benchmarking.process_benchmark_results import _normalize
+from estimagic.benchmarking.process_benchmark_results import create_performance_df
 
 PROBLEMS = ["prob1", "prob2", "prob3"]
